requirements.txt
Django==5.2.4
djangorestframework==3.16.0
celery==5.5.3
redis==6.2.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-dotenv==1.1.1
djangorestframework-simplejwt==5.5.1
smsaero-api==3.0.0